Increased initial rewritten code mem alloc, removed some asserts
I decided to increase the initial size of the allocated memory for
rewritten code to be just big enough so that it didn't require any
expansion in some quick tests.  In general, if the pattern holds
(so it almost never needs to increase the allocated size), this
should provide a small speed increase and less memory usage.

I also conditionally removed some asserts that should only have
cost a little bit of runtime performance and were only useful for
verifying that things were working.  I can restore the asserts if
I encounter problems.

In practice, this really didn't result in a very noticeable
improvement in performance, but they were easy changes.  Now it's
time to try to tackle the disassembly overhead, if possible.