Changed chunk size to be statically defined
For some reason I don't recall, I originally made the chunk size
a parameter to pass to the gen_code function, allowing the chunk
size to vary between different rewritten regions.  However, if it
were actually changed, jumps between regions would lead to crashes.
I also can't think of a reason that this value would need to be
changed at runtime.  Therefore, it is simpler and more efficient to
make it a compile-time constant.

While the code that refers to chunk size was called numerous times
throughout the rewriter, this change did not lead to a notable
speed improvement, although it seems to have subtly improved it.
The change was worth doing regardless, and it is a good time to
tidy up the code while searching for optimizations.