Factored out calls to is_target to de-duplicate calls
The application-defined is_target call was being called every time
that gen_padding and check_target were being called, so it was
being called at least twice for every instruction.  I changed it
so that I call is_target before both of those other functions and
then pass the result to those functions.  Right now is_target is
just a function that returns true (for testing), so the performance
improvement is basically non-existent.  However, if this function
becomes more complicated, then this change might make a bigger
difference.  Regardless of if it ever yields better performance,
I think this is a better approach.