Added translate_address function, handle another call encoding
Any code that calls into jit-compiled code can now optionally
have a call to translate_address inserted at that call point.
Doing this results in a slight performance improvement, as we then
don't entirely rely on the segfault handler to transition into
executing the rewritten code.  However, this does require some
modifications to the original code, and it doesn't address the
issue of jit code calling back into the original code.  When the
original code returns to the jit code, the segfault handler is
used.  Therefore, I need to check again for why I switched to
pushing old return addresses to the stack instead of just using
call instructions.  There was some compatibility issue, but I don't
recall the specifics.

I also added code to handle another type of encoding of call
instructions in miniverse_spatcher.py.  I encountered a type of
call instruction that was 4 bytes long due to an additional SIB
byte, forced by using register $esp for the target address.  This
did actually appear in the original code, so this has fixed that
particular issue.