Added optimization for certain direct calls, added new tests
Even after I added code to translate the addresses for jumps from
the original program into jit code (avoiding the segfault handler),
I still found a large number of jumps triggering segfaults.  To
optimize performance, we want to minimize using the segfault
handler, since it is very inefficient.

I found that some of the segfaults were triggered by original code
returning to the jit code: the jit code calls original code, which
then returns when it's done.  However, the jit code by default
always pushed the old address on the stack for rewritten calls,
so the return in the original code tried to return to the old jit
code, passing through the segfault handler.  I changed it so that
direct calls to an address outside of jit code will remain as
call instructions instead of a "push old address + jump" combo.
However, this only improved performance somewhat, and there were
still a lot of segfaults.

Many segfaults were still happening for calls between jit regions.
A call between regions can't really do a lookup at the time of
rewriting, because regions are rewritten multiple times and the
address could become stale.  However, what I was doing before was
simply jumping to the old address, again forcing a lookup via the
segfault handler.  I changed it so that it instead rewrites a
direct call between jit regions to be a rewritten indirect call,
which means that the lookup is done before jumping as the rewritten
jit code runs.  This is slower than an ordinary direct call, but
is definitely faster than a segfault.  Performance improved even
more than before.

I also added several new tests.  One, bigmem, tests for allocating
a large amount of jit code, and the other two are for testing the
code related to indirect calls using the stack pointer and for
testing cross-jit-region calls.