Added optimization for certain direct jumps, updated test
Similarly to direct calls that called code in a separate jit
region, jumps to a separate jit region also resulted in a slow
fallback to the segfault handler lookup.  I refactored the direct
call/jump code to better handle this case for both calls and jumps,
and the result is noticeably better performance!

My changes at first resulted in some very difficult to track down
crashes, as the newly generated jump instructions were not marking
themselves as jump targets (even though ALL instructions are
currently marked as jump targets).  This resulted in very specific
cases in which code that tried to jump to a direct jump instruction
would actually jump to the instruction before it, which can lead to
all sorts of confusing misbehavior (in this case, nondeterministic
crashes).  I fixed it and added a case to test15 to test both
cross-region direct jumps in general and this specific bug.