Fixed some compatability bugs with luajit
I had made a lot of changes since I first got miniverse working
with luajit, and I had to make some small changes to get it to work
on the benchmarks I previously tested with.  Performance is worse,
mostly from having to switch to a permissive "all addresses are
targets" policy for full compatibility.  With some additional
tests, it should be possible to get a much less permissive policy
with much better performance, but that is going to have to wait for
later.

These changes haven't been tested to work with firefox, so I will
need to go back and some point to make sure it didn't break
compatibility for some reason.

These changes were made a few weeks back when I got luajit working,
but I had left a few miscellaneous changes uncommitted as I tested
things.