Started implementation of "standalone" miniverse binary
As miniverse is implemented now, it must be compiled into a project
as a library, but I am working to get it so that miniverse can be
loaded as a binary blob either when rewriting a binary or mapped
into the memory space at runtime.  This will hopefully help with
getting it working with a binary unpacker such as UPX.  However,
this approach does depend on it being possible to run miniverse
code without it ever being touched by the loader, so it is compiled
as a position-independent executable.  The current test I have
written, in test-standalone.c, doesn't work yet, as I'm not mapping
the data segment with an alignment that mmap will accept yet. After
that, though, I'm hoping I can confirm that miniverse can function
correctly even if it is not compiled in with a project.

One additional problem to solve is with the wrappers around mmap
and mprotect.  For my test, I will "wrap" them manually, but for
an unpacker I may manually replace all calls to mmap/mprotect in
the unpacker code with calls to the wrapper.  This wouldn't
provide protection for mmap/mprotect calls made by the unpacked
binary itself, but that might be a separate problem to solve.