Got standalone test working
Now it is possible to load miniverse into an application without
relying on the loader for anything.  This standalone version will
make it easier to embed miniverse into a binary, which should help
with integrating it with a binary packer.

I had to make some changes to avoid any functions that use dynamic
memory allocations, because we don't want to interfere with the
original binary's mallocs.  There were a couple functions I had to
write bespoke replacements for, and they only implement the subset
of functionality I needed.

The standalone test is very fragile, because the addresses of
functions are hard-coded, and any changes to miniverse may change
them.  Also, while the standalone binary is compiled as
position-independent, it must be loaded at a specific address.
Since we aren't running the loader, it isn't patching relocations.
The udis86 disassembler has large tables of (non-relative)
addresses pointing to other tables, and I would need to implement
some subset of loader functionality to patch those tables before
it could be loaded at arbitrary load-time addresses.