Adjusted addresses for standalone test
The addresses and offsets for the generated standalone binary are
so fragile that compiling it on two separate instances of Ubuntu
20.04 resulted in two slightly different binaries.  A more stable
solution would be to derive the addresses from the generated
standalone binary, but for testing I'm still going to set them
manually.  I have changed the values for the machine I will be
testing the binary packer on.