Added assembly for loading standalone miniverse
Now the standalone test loads miniverse in with a small,
handwritten bit of relocatable assembly.  This can be slotted in
to the entry stub assembly used in UPX, the binary packer I'm
testing with.  There is a tiny amount of assembly code that is
not compressed when a compressed UPX binary is launched.  Adding
this will allow my code to load miniverse like a strange, isolated
shared library that doesn't interact with the rest of the program.
The loaded miniverse will not have any wrappers around mmap or
mprotect, so I will need to manually wrap all calls made by UPX.