Standalone use custom libc, custom replacements for some lib calls
When running the standalone version of miniverse in UPX, the normal
C runtime environment has not been set up.  A normal C program
includes some hint given to the loader that normally initializes
the gs register prior to the binary entry point.  However, it is
not set at all in UPX.  Unfortunately, code using the gs register
is automatically inserted by the compiler into many functions to
retrieve a thread-local stack canary value.  These references lead
to a crash if gs is unset.  By compiling a custom version of libc
with the -fno-stack-protector flag, I can avoid this problem.

I also wrote custom replacements for several library functions that
used the gs register, but for some other thread-local variables if
I recall correctly.  Some of the changes made in this commit were
made weeks back, so I'm not sure about the details.

I also wrote a script that parses the output from readelf on
standalone to get the correct addresses that need to be entered
into test-standalone.S and test-standalone.c.  Doing it manually
was annoying and tedious.

As of this commit, I still do not actually have miniverse working
completely in UPX.  I'm most of the way there, and I was going to
wait to commit my changes until I had it working, but I had to
pause this work for another project.  Since I had been away from
the project for a while, I had to spend some time catching up with
where I had been before pausing, and I decided it was better to
commit these changes before making new changes in this repo.