Added func addresses to standalone script, removed debug printfs
Now get_standalone_offsets.py also shows the locations of the
register_handler, __wrap_mmap, and __wrap_mprotect functions in
the standalone binary.  Now the script provides all the information
needed to update the addresses and values in test-standalone.c and
test-standalone.S.  Not all addresses have been updated in that
test, as I have only been testing the assembly implementation.

Several debug printfs were causing problems because they include
instructions that use the gs segment register.  Printf tries to
retrieve errno, which is thread-local, and puts (which I think is
automatically internally called instead of printf when no extra
arguments are passed to printf) refers to gs in _IO_acquire_lock.
Since I don't really NEED to print out anything in the standalone
version, I made it so that these calls to printf are only inserted
when DEBUG is defined.

These changes also changed the layout of standalone, so I have
updated the test accordingly.