Handle jmps to jit code with valid & wrong address on top of stack
I previously added support for cases when the main binary jumped
into jit code rather than calling into it, but I assumed that when
my handler checked for a return address, the resulting "address"
when jumping to the jit code would result in a garbage value that
didn't point into a mapped memory region.  However, UPX leaves an
address to a valid memory region at the top of the stack that is
NOT the region that it was calling from, leading to that code
segment not being mirrored, and instead some other code segment
being mirrored instead.  That meant that a later call back into
the original binary from the jit code performed a lookup at a
fixed offset from that original code, and nothing would be mapped
there, leading to a segfault.  I fixed this by mirroring up to two
regions: the caller's region, as determined by whichever address
is at the top of the stack, and the first executable region in
memory.  This will heuristically work...probably in a lot of cases,
but it's still a heuristic.  If those two regions are identical,
then I only mirror once.

I have encountered another problem with the allocation of the
memory at the fixed offset.  With UPX, the initial extracted
unpacking code is at a very different address from the final
extracted binary address, and the large offset buffer I allocate
doesn't end up covering it.  However, it seems that it fortunately
doesn't allocate memory in the location that I would need if I
were to allocate an additional buffer for the extracted binary.
Unfortunately, with ASLR and outside of a debugger, I can't rely
on that always being the case.  My current approach is to attempt
to allocate a big buffer up front based on the first address I
need to rewrite, and I guarantee that my fixed offset works with
that allocation.  If I can't find *SOME* suitable offset, I will
exit.  Then I hope that all jit regions are clustered together and
most new jit code is mapped after the initial jit region.  However,
finding out that jit code was allocated at some wacky address way
later into execution can sabotage that plan.  I simply don't have
a working solution to map lookup tables in memory at arbitrary
locations and index them quickly and universally; all jit code
needs to know how to find the lookup table.  I could use a tiered
system, with a table in a fixed location that gives mappings for
every jit region's lookup tables, but that is going to be way
slower than adding a simple offset.  For now, I think I'll just
write code to detect if a region I'm about to rewrite for the
first time doesn't have anything mapped at the fixed offset I
picked, and attempt to allocate it if not.  If there is already
some region mapped that collides with those addresses, then it
will fail.  For now, I'll just hope it doesn't fail.

I also fixed the comment I had written for the loope instructions,
which had some inaccurate example assembly.