Fixed issue with jit regions being allocated at distant addresses
My initial approach to allocating the fixed offset buffer was to
allocate a much larger region than the initial jit region and
hoping that most subsequent jit region allocations would be in a
range after the initial allocation.  However, UPX allocates two
unrelated code regions that both need to be rewritten, yet are
allocated at very different addresses.  I solve this by checking
right before I rewrite a region for the first time whether its
fixed offset region falls into the large region I initially
allocated.  If not, I try to allocate a region just for that
region.  If there is already something allocated in that address
range, it will fail, but this solution appears to work for UPX.

I have encountered yet another problem, though, in which the
lengths UPX passes to mmap and mprotect are different, and neither
is page aligned.  The mmap/mprotect call automatically aligns the
length to the nearest boundary, but I don't.  This discrepancy
seems to have led to me not rewriting some crucial instructions.