Align mmap and mprotect lengths to page boundaries
I found that UPX passed a length to mmap that was a few bytes
longer than a subsequent call to mprotect.  It wasn't clear why,
but the result during normal execution isn't impacted because
permissions are at a page granularity.  However, I took the lengths
verbatim in my mmap and mprotect wrappers, and it meant I didn't
actually rewrite the entirety of the allocated memory in those
regions.  Normally, bytes outside of the length passed to mmap
or mprotect are not intended to be executed, but for some reason
the last few bytes in a region allocated by UPX were crucial but
not included in the length passed to the mprotect call adding
executable permissions.  I have fixed this by rounding up to the
normal page size.  I assume large pages aren't being used.

Unfortunately, I have encountered yet another issue that I haven't
solved yet.  For some reason, it seems like the displacements in
some call instructions are different when loaded by UPX than they
were in the original uncompressed binary, and I'm not sure what's
causing it.