Support segment override prefix in jmp/call, map table for VDSO
Now jmp and call instructions have limited support for segment
override prefixes.  I should be able to handle any of the segment
prefixes (although I'm currently only testing gs, ds, and cs), but
I still expect a single prefix and don't support other prefixes.
Also, while I should be able to handle segment override prefixes
for direct calls/jmps, I simply omit the prefix, since it doesn't
really make sense to have a segment override prefix on an
instruction that doesn't reference memory.  I haven't tested this
particular instruction variant either, so I don't know if there
are bugs in the implementation, but I think it should never show
up in normal code.  If I DID encounter an example of it, or if the
codebase was being prepared for production use, I would need to
add tests for this.  For now, I just want to make a note that this
could be something to keep in mind for later.

I also found that calls to the VDSO (such as by rewritten code that
wants to make a system call) didn't work reliably since I hadn't
mirrored the VDSO.  Fortunately, the __kernel_vsyscall function in
the VDSO is 16-byte aligned on my machine, but the initial byte
of the function would make miniverse think that it needs
to perform a lookup if I just mirrored the VDSO.  I realized that
mirroring memory segments like I have been doing is not needed,
and that instead I needed to just allocate memory for the lookup
tables with empty entries that indicate no lookup should be done
for any address in that region.  Therefore, I changed the code to
do this instead of mirroring.  This worked for all the regions
I originally mirrored, and it should work in general.  A
non-rewritten region simply can have an empty table, with no need
to mirror it.  This means that technically there is no longer a
need to insert nops at the start of functions in non-rewritten
regions like I'm currently doing, although the functions still
will need to be 16-byte aligned.