Fixed bug parsing /proc/self/maps, avoid trying to map invalid addr
When I was parsing /proc/self/maps to check for specific segments,
I incorrectly tried to parse after the last line in the file if an
address was not found in any segment.  I was getting the contents
of the file one line at a time, but I assumed the result of the
call to file_get_line always had a populated line.  When I return
false from file_get_line, that meant that it reached EOF.  I still
assumed that there was something in the last line, and that it was
terminated by the end of the file, but the last line in the file
actually ends in a newline (leaving zero bytes after the newline
and before EOF).  This resulted in it trying to parse garbage bytes
after the null-terminated zero byte string (actually 1 byte due to
another bug I also fixed).  Now, it will exit as soon as the
function file_get_line return false, since I know /proc/self/maps's
last line ends with a newline.

I also added a check so that I don't try to map a fixed-offset
table for the calling address in set_fixed_offset if the calling
address is actually an invalid address not present in any mapped
memory region.  This can happen if a program jumps to JIT code
instead of calling into it, as I make a guess that the address at
the top of the stack is a return address.  If it isn't, then the
value can be anything, and not checking whether the value is a
valid address led to me trying to map an offset from an arbitrary
non-address value.  This could lead to a failure if something is
already at that address.  I previously had changed the "caller
segment" address to the first executable segment's address if the
caller address was invalid, which avoided the problem but was
redundant, as I always map the table for the first executable
segment regardless.  Therefore, I removed a redundant call to
find_first_exec_segment, but this revealed a new issue with
mapping a table for an invalid region, which I fixed.
In short, I removed a redundant operation and added a less
expensive extra check to compensate.