Cleaned up code that parses /proc/self/maps
I was actually opening, parsing, and closing /proc/self/maps about
10 times when starting up miniverse.  This led to a lot of
redundant parsing code, since 5 separate functions contained
similar or identical functionality for parsing the file.  This
duplication of code and of effort via unnecessary parsing bothered
me, so I cleaned it up so that the parsing only happens twice.  The
two points at which the memory maps are needed happen far enough
apart that I wanted to have an up-to-date memory map, so I cache
the contents of /proc/self/maps before each of those two points.

Despite being theoretically more time efficient, there is no
measurable improvement in performance from my tests, so the main
benefit is that the code is cleaner with less duplicated
functionality.  A minor downside is that my cache of the memory
maps requires a newly allocated region of memory.  I allocate 4
pages (16384 bytes), which is enough for 1024 16-byte entries.
I checked the number of lines in the maps file for the Firefox
instance running on my machine, and it was a little over 600, so
1024 should be enough for very large software.  However, it is
still a new restriction that wasn't present before, and I didn't
implement expanding the allocated size dynamically.  It should
force a crash with an error message if there are too many mapped
regions, so the size could be expanded if I ever encounter a
binary that triggers that crash.  I also only allocate 7 bytes for
a region's name, so if I ever need that data for something other
than finding the "[vdso]" string, then I would need to rethink
the current implementation.  I'm also assuming that there is no
padding in the new maps_region_t struct I added, so each entry
takes up exactly 16 bytes.  This assumption also relies on 4-byte
addresses, so it wouldn't work as-is in a 64-bit process.