Added optimization for calls to __kernel_vsyscall
Calls are made to the __kernel_vsyscall function for performing
fast system calls in x86-32 binaries.  These calls were converted
to a push/jump pair by miniverse, leading to a segfault when the
function returned to an old address.  This worked, but led to
additional segfaults whenever an application performed a fast
system call.  I therefore added an optional optimization that
doesn't rewrite calls to this specific function.

This does reduce segfaults as expected, but does not result in
an expected performance improvement.  The rewritten code performs
more poorly than expected.