Switched to multi-byte nops, aligned padded instructions
I decided to test the performance of chunk-aligning instructions
and compressing the sequences of multiple nops.  These changes
improved performance more than the previous optimization, although
still not as much as I would like.

Previously, instructions identified as jump targets were aligned
one byte into a chunk, as the first byte was a marker indicating
that the address is a jump target.  However, this has been obsolete
since I changed to using a fixed offset, as the marker is now in
the entry at the fixed offset.  This alignment resulted in a minor
improvement to performance.

I originally added padding by placing as many single-byte 0x90 nop
instructions as were needed.  I replaced that with multi-byte nops
instead, and it improved performance more than aligning
instructions did.