Removed special-case padding for call instructions
In the gen_padding function, I padded all call instructions to the
end of a chunk instead of to the start of a chunk so that a return
from a call instruction returned directly to a chunk-aligned
address.  However, I felt that this could be done more efficiently
by ensuring that the last set of instructions for a rewritten call
is exactly 16 bytes by padding the sequence ahead of time.  That
way, the sequence is automatically padded to the end of the chunk
without having to perform two checks and insert extra padding like
it did before.

This change has worked with everything I tested so far.  However,
I'm suspicious that direct calls may not be padded correctly after
this change, including direct calls within the same JIT region or
calls from a JIT region to a non-JIT region.  I want to
double-check this and ensure there is no regression.  Again, all
tests work fine, so if there is an issue, I may want to modify my
tests to catch it.  I didn't notice this potential issue until
after I ran a number of tests, so I want to save what I have now
before performing further checks.