Fixed bug with memory range validity check
When I checked whether a memory range was valid (used when checking
if a specific fixed offset would result in any memory collisions),
I had accidentally omitted a case in which the memory range
completely enclosed another region as opposed to only partially
overlapping it or being within it.  This resulted in some rare
cases in which a fixed offset was chosen that interfered with an
existing memory mapping.