Added optional shadow stack implementation
Now miniverse can be optionally compiled to implement an
overwriting, no-zeroing, parallel shadow stack.  This is the same
type of shadow stack that I implemented in multiverse, although
there are some implementation differences.

The shadow stack normally would not work without some preparation
in the target application, as the shadow stack is only implemented
for the jit code.  This means the original generating code that
calls into the jit code will not have a shadow stack entry for that
initial call, so any returns from the jit code back into the
application code would fail.  I address this by secretly writing a
shadow stack entry when a segfault occurs, thereby adding an entry
every time the original code tries to call into the original jit
code.  This currently inserts an entry regardless of whether the
the segfault is caused by a call instruction, but it works for a
proof-of-concept.  In real usage the main application and the jit
code should both use a shadow stack.

Any optimizations that perform a lookup to avoid the segfault will
fail.  Also, UPX doesn't work with the shadow stack because it
intentionally uses a mismatched return in its escape hatch.  Some
kind of special case would need to be implemented for that policy
violation.