Increased size of nanoseconds field in stat timers, print as signed
When measuring the timing of each component in miniverse, I was
storing the total time in timespec structs.  Due to
the way that I am measuring tiny time slices, it is possible
for the nanoseconds field to be negative to compensate for the
seconds field rolling over (if a few ms pass, but the seconds
field ticks over, then the time duration is 1 second and negative
millions of nanoseconds).  This should work fine while measuring,
but a long in a 32-bit binary can only hold a little over 2
billion, which is only a little over 2 seconds when storing
nanoseconds.  It might be possible that the nanoseconds field would
overflow or underflow when tallying up many tiny timeslices, so I
increased the field to a long long, which should be plenty large.

However, the bigger issue was the fact that I was printing the
nanoseconds field as an unsigned integer.  This meant that if the
field was negative, I was printing the number of nanoseconds as
positive, which skewed some results by a few seconds.  I ran
some tests and found that the times are still similar to what
they were before, but now with more consistency, as a negative
number of nanoseconds no longer skews the results.  Whether the
nanoseconds value for a particular timer is positive or negative
appears to be pretty random, depending on how many times the
measurement takes place across a second boundary, so running
multiple times for my previous tests reduced the impact of the
outliers caused by the wrong numbers.