Added option for traditional shadow stack
Now miniverse can insert a traditional shadow stack or a parallel
shadow stack.  Due to the nature of each type of shadow stack,
some of the tests I wrote will fail for each type.  This is due to
intentional policy violations in the tests, in which calls and
returns do not match.  Jit code that violates the policy of a
shadow stack will also fail.