Doubled space allocated for shadow stack
Since I don't expand the shadow stack after first allocating it,
it's possible for an application with heavy stack usage to expand
the stack enough that it will exceed the original stack's size and
allocate more stack space.  Doubling the initial shadow stack size
appears to be enough to avoid issues with the applications I am
testing.  Something more robust would be needed for a
production-grade solution.